#33: Split up interpreter tests
