+++ agent/src/main/kotlin/com/octogonapus/omj/agent/interpreter/Interpreter.kt
@@ -1,36 +1,48 @@
+/*
+ * This file is part of OMJ.
+ *
+ * OMJ is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * OMJ is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with OMJ.  If not, see <https://www.gnu.org/licenses/>.
+ */
 package com.octogonapus.omj.agent.interpreter
 
-import org.objectweb.asm.Opcodes.ACONST_NULL
-import org.objectweb.asm.Opcodes.BIPUSH
-import org.objectweb.asm.Opcodes.DCONST_0
-import org.objectweb.asm.Opcodes.DCONST_1
-import org.objectweb.asm.Opcodes.FALOAD
-import org.objectweb.asm.Opcodes.FCONST_0
-import org.objectweb.asm.Opcodes.FCONST_2
-import org.objectweb.asm.Opcodes.IALOAD
-import org.objectweb.asm.Opcodes.IASTORE
-import org.objectweb.asm.tree.AbstractInsnNode
-import org.objectweb.asm.tree.InsnList
-import org.objectweb.asm.Opcodes.ICONST_5
-import org.objectweb.asm.Opcodes.ICONST_M1
-import org.objectweb.asm.Opcodes.LALOAD
-import org.objectweb.asm.Opcodes.FASTORE
-import org.objectweb.asm.Opcodes.DALOAD
-import org.objectweb.asm.Opcodes.DASTORE
 import org.objectweb.asm.Opcodes.AALOAD
 import org.objectweb.asm.Opcodes.AASTORE
+import org.objectweb.asm.Opcodes.ACONST_NULL
 import org.objectweb.asm.Opcodes.BALOAD
 import org.objectweb.asm.Opcodes.BASTORE
+import org.objectweb.asm.Opcodes.BIPUSH
 import org.objectweb.asm.Opcodes.CALOAD
 import org.objectweb.asm.Opcodes.CASTORE
+import org.objectweb.asm.Opcodes.DALOAD
+import org.objectweb.asm.Opcodes.DASTORE
+import org.objectweb.asm.Opcodes.DCONST_0
+import org.objectweb.asm.Opcodes.DCONST_1
 import org.objectweb.asm.Opcodes.DUP
 import org.objectweb.asm.Opcodes.DUP2
 import org.objectweb.asm.Opcodes.DUP2_X1
 import org.objectweb.asm.Opcodes.DUP2_X2
 import org.objectweb.asm.Opcodes.DUP_X1
 import org.objectweb.asm.Opcodes.DUP_X2
-import org.objectweb.asm.Opcodes.SALOAD
-import org.objectweb.asm.Opcodes.SASTORE
+import org.objectweb.asm.Opcodes.FALOAD
+import org.objectweb.asm.Opcodes.FASTORE
+import org.objectweb.asm.Opcodes.FCONST_0
+import org.objectweb.asm.Opcodes.FCONST_2
+import org.objectweb.asm.Opcodes.IALOAD
+import org.objectweb.asm.Opcodes.IASTORE
+import org.objectweb.asm.Opcodes.ICONST_5
+import org.objectweb.asm.Opcodes.ICONST_M1
+import org.objectweb.asm.Opcodes.LALOAD
 import org.objectweb.asm.Opcodes.LASTORE
 import org.objectweb.asm.Opcodes.LCONST_0
 import org.objectweb.asm.Opcodes.LCONST_1
@@ -38,12 +50,16 @@ import org.objectweb.asm.Opcodes.NEWARRAY
 import org.objectweb.asm.Opcodes.NOP
 import org.objectweb.asm.Opcodes.POP
 import org.objectweb.asm.Opcodes.POP2
+import org.objectweb.asm.Opcodes.SALOAD
+import org.objectweb.asm.Opcodes.SASTORE
 import org.objectweb.asm.Opcodes.SIPUSH
+import org.objectweb.asm.tree.AbstractInsnNode
+import org.objectweb.asm.tree.InsnList
 import org.objectweb.asm.tree.InsnNode
 import org.objectweb.asm.tree.IntInsnNode
 
 internal class Interpreter(
-        private val insnList: InsnList
+    private val insnList: InsnList
 ) {
 
     private val stackMap: MutableMap<AbstractInsnNode, OperandStack> = HashMap()
@@ -98,10 +114,14 @@ internal class Interpreter(
             is InsnNode -> when (insn.opcode) {
                 NOP -> OperandStackOperation.NOP
                 ACONST_NULL -> OperandStackOperation.PushNullRef
-                in ICONST_M1..ICONST_5 -> OperandStackOperation.PushConstInt((insn.opcode - ICONST_M1) - 1)
-                in LCONST_0..LCONST_1 -> OperandStackOperation.PushConstLong(insn.opcode - LCONST_0.toLong())
-                in FCONST_0..FCONST_2 -> OperandStackOperation.PushConstFloat(insn.opcode - FCONST_0.toFloat())
-                in DCONST_0..DCONST_1 -> OperandStackOperation.PushConstDouble(insn.opcode - DCONST_0.toDouble())
+                in ICONST_M1..ICONST_5 ->
+                    OperandStackOperation.PushConstInt((insn.opcode - ICONST_M1) - 1)
+                in LCONST_0..LCONST_1 ->
+                    OperandStackOperation.PushConstLong(insn.opcode - LCONST_0.toLong())
+                in FCONST_0..FCONST_2 ->
+                    OperandStackOperation.PushConstFloat(insn.opcode - FCONST_0.toFloat())
+                in DCONST_0..DCONST_1 ->
+                    OperandStackOperation.PushConstDouble(insn.opcode - DCONST_0.toDouble())
                 IALOAD -> OperandStackOperation.LoadIntFromArray
                 IASTORE -> OperandStackOperation.StoreIntoIntArray
                 LALOAD -> OperandStackOperation.LoadLongFromArray
