+++ agent/src/main/kotlin/com/octogonapus/omj/agent/interpreter/OperandUtil.kt
@@ -1,31 +1,37 @@
+/*
+ * This file is part of OMJ.
+ *
+ * OMJ is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * OMJ is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with OMJ.  If not, see <https://www.gnu.org/licenses/>.
+ */
 package com.octogonapus.omj.agent.interpreter
 
 import org.objectweb.asm.Opcodes.AALOAD
 import org.objectweb.asm.Opcodes.AASTORE
-import org.objectweb.asm.Opcodes.ACONST_NULL
 import org.objectweb.asm.Opcodes.BALOAD
 import org.objectweb.asm.Opcodes.BASTORE
-import org.objectweb.asm.Opcodes.BIPUSH
 import org.objectweb.asm.Opcodes.CALOAD
 import org.objectweb.asm.Opcodes.CASTORE
 import org.objectweb.asm.Opcodes.DALOAD
 import org.objectweb.asm.Opcodes.DASTORE
-import org.objectweb.asm.Opcodes.DCONST_0
 import org.objectweb.asm.Opcodes.FALOAD
 import org.objectweb.asm.Opcodes.FASTORE
-import org.objectweb.asm.Opcodes.FCONST_0
 import org.objectweb.asm.Opcodes.IALOAD
 import org.objectweb.asm.Opcodes.IASTORE
-import org.objectweb.asm.Opcodes.ICONST_M1
 import org.objectweb.asm.Opcodes.LALOAD
 import org.objectweb.asm.Opcodes.LASTORE
-import org.objectweb.asm.Opcodes.LCONST_0
-import org.objectweb.asm.Opcodes.NEWARRAY
-import org.objectweb.asm.Opcodes.NOP
 import org.objectweb.asm.Opcodes.SALOAD
 import org.objectweb.asm.Opcodes.SASTORE
-import org.objectweb.asm.Opcodes.SIPUSH
-import org.objectweb.asm.Opcodes.T_BOOLEAN
 import org.objectweb.asm.Opcodes.T_BYTE
 import org.objectweb.asm.Opcodes.T_CHAR
 import org.objectweb.asm.Opcodes.T_DOUBLE
