+++ testUtil/src/main/kotlin/com/octogonapus/omj/testutil/KotestUtil.kt
@@ -21,6 +21,7 @@ import io.kotest.matchers.MatcherResult
 import io.kotest.matchers.neverNullMatcher
 import io.kotest.matchers.should
 import io.kotest.matchers.shouldNot
+import io.kotest.property.Exhaustive
 
 fun <T> Array<T>.shouldHaveInOrder(vararg ps: (T) -> Boolean) =
     asList().shouldHaveInOrder(ps.toList())
@@ -121,4 +122,10 @@ fun <T> hasExactlyInOrder(predicates: List<(T) -> Boolean>): Matcher<Collection<
             { "$actual did not match the predicates $predicates in order" },
             { "$actual should not match the predicates $predicates in order" }
         )
+    }
+
+fun <A, B : A, C : A> Exhaustive<B>.merge(other: Exhaustive<C>): Exhaustive<A> =
+    object : Exhaustive<A>() {
+        override val values: List<A> = this@merge.values.zip(other.values)
+            .flatMap { listOf(it.first, it.second) }
     }
